requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
argcomplete==3.6.2
asyncio==3.4.3
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.38.25
botocore==1.38.25
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
cryptography==45.0.3
dataclasses==0.6
DateTime==5.5
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fasta2a==0.2.11
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-genai==1.16.1
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
junos-eznc==2.7.4
langchain-core==0.3.62
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.42
logfire==3.16.1
logfire-api==3.16.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.1
mdurl==0.1.2
mistralai==1.8.1
napalm==5.0.0
narwhals==1.41.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.5.0
netutils==1.13.0
ntc_templates==7.9.0
numpy==2.2.6
openai==1.82.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pathlib==1.0.1
pillow==11.2.1
prompt_toolkit==3.0.51
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-ai==0.2.11
pydantic-ai-slim==0.2.11
pydantic-evals==0.2.11
pydantic-graph==0.2.11
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyeapi==1.0.4
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
scp==0.15.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
streamlit==1.45.1
streamlit-autorefresh==1.0.1
tenacity==9.1.2
textfsm==1.1.3
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
transitions==0.9.2
ttp==0.9.5
ttp-templates==0.3.7
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yamlordereddictloader==0.4.2
zipp==3.22.0
zope.interface==7.2
zstandard==0.23.0
